requirements.txt
openai>=1.0.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.15.0
python-dotenv>=1.0.0
jupyter>=1.0.0
notebook>=6.5.0
ipywidgets>=8.0.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
httpx>=0.25.0
aiofiles>=23.0.0
# Database dependencies
sqlalchemy>=2.0.0
pymysql>=1.0.0
cryptography>=41.0.0
# Analytics and ML dependencies
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
tqdm>=4.65.0
aiohttp>=3.8.0
playwright>=1.49.0
# Semantic labelling (intercom_conversation_taxonomy.py --labelling-method semantic)
transformers>=4.30.0
torch>=2.0.0